ml_models/requirements.txt
pymongo==4.6.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
